Extended contacts app

+++ Practice/contacts/contacts_app/main_window.py
@@ -1,4 +1,5 @@
-from PyQt6.QtGui import QStandardItem, QStandardItemModel
+from PyQt6.QtCore import Qt
+from PyQt6.QtGui import QStandardItem, QStandardItemModel, QAction, QIcon
 from PyQt6.QtWidgets import (
     QApplication, QMainWindow, QPushButton, QListView, QGroupBox, QLineEdit, QTextEdit, QLabel,
     QHBoxLayout, QVBoxLayout, QWidget, QMessageBox
@@ -50,14 +51,19 @@ class MainWindow(QMainWindow):
 
         # Buttons
         self.quit_btn = QPushButton("Quit", self)
+        self.quit_btn.setHidden(True)
+        
         self.delete_btn = QPushButton("Delete", self)
+        self.delete_btn.setDisabled(True)
+
         self.save_btn = QPushButton("Save", self)
 
         # Buttons layout
         buttons_layout = QHBoxLayout()
-        buttons_layout.addWidget(self.save_btn)
         buttons_layout.addWidget(self.delete_btn)
-        buttons_layout.addWidget(self.quit_btn)
+        buttons_layout.addStretch()
+        buttons_layout.addWidget(self.save_btn)
+        # buttons_layout.addWidget(self.quit_btn)
 
         # Right column layout
         right_column = QVBoxLayout()
@@ -74,15 +80,39 @@ class MainWindow(QMainWindow):
         main_widget.setLayout(main_layout)
         self.setCentralWidget(main_widget)
 
+        # Toolbars and actions
+        main_toolbar = self.addToolBar("Main")
+        main_toolbar.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextUnderIcon)
+
+        self.new_contact_action = QAction(QIcon("contacts_app/icons/add.png"), "Add contact", self)
+        self.new_contact_action.setIconText("Add contact")
+        self.new_contact_action.setToolTip("Add a new contact")
+
+        self.quit_action = QAction(QIcon("contacts_app/icons/logout.png"), "Quit", self)
+        self.quit_action.setIconText("Quit")
+
+        main_toolbar.addAction(self.new_contact_action)
+        main_toolbar.addSeparator()
+        main_toolbar.addAction(self.quit_action)
+
+        # Menubar
+        main_menu = self.menuBar().addMenu("Contacts")
+        main_menu.addAction(self.new_contact_action)
+        main_menu.addSeparator()
+        main_menu.addAction(self.quit_action)
+
         self.resize(800, 600)
         self.setWindowTitle("My Contacts")
 
     def initActions(self):
         self.quit_btn.clicked.connect(self.quit)
+        self.quit_action.triggered.connect(self.quit)
+
         self.save_btn.clicked.connect(self.saveClicked)
         self.delete_btn.clicked.connect(self.deleteClicked)
         self.contacts_list.activated.connect(self.contactActivated)
         self.contacts_list.clicked.connect(self.contactActivated)
+        self.new_contact_action.triggered.connect(self.newContactActivated)
 
     def refreshContacts(self):
         contacts = self.contacts_service.get_contacts()
@@ -96,6 +126,13 @@ class MainWindow(QMainWindow):
 
             self.contacts_model.appendRow([name, email])
 
+    def newContactActivated(self):
+        self.current_contact = None
+        self.contacts_list.clearSelection()
+        self.clearFields()
+        self.delete_btn.setDisabled(True)
+        self.name_edit.setFocus()
+
     def contactActivated(self, index):
         contact_id = self.contacts_model.itemFromIndex(index).data()
         self.current_contact = self.contacts_service.get_contact(contact_id)
@@ -105,25 +142,43 @@ class MainWindow(QMainWindow):
         self.address_edit.setText(self.current_contact["address"])
         self.notes_edit.setMarkdown(self.current_contact["notes"])
 
+        self.delete_btn.setDisabled(False)
+
     def saveClicked(self):
+        is_edit = False if self.current_contact is None else True
+
+        if not is_edit:
+            self.current_contact = {}
+
         self.current_contact["name"] = self.name_edit.text()
         self.current_contact["email"] = self.email_edit.text()
         self.current_contact["address"] = self.address_edit.text()
         self.current_contact["notes"] = self.notes_edit.toMarkdown()
 
-        self.contacts_service.update(self.current_contact)
+        if is_edit:
+            self.contacts_service.update(self.current_contact)
+            QMessageBox.information(self, "Updated", "Your contact has been successfully updated")
+        else:
+            self.contacts_service.create(self.current_contact)
+            QMessageBox.information(self, "Created", "Your contact has been successfully created")
+
         self.refreshContacts()
         
-        QMessageBox.information(self, "Updated", "Your contact has been successfully updated")
     
     def deleteClicked(self):
+        response = QMessageBox.question(
+            self, 
+            "Delete contact", 
+            "Dou you really want to delete {} contact".format(self.current_contact["name"])
+        )
+
+        if response != QMessageBox.StandardButton.Yes:
+            return
+
         self.contacts_service.delete(self.current_contact)
         self.current_contact = None
 
-        self.name_edit.clear()
-        self.email_edit.clear()
-        self.address_edit.clear()
-        self.notes_edit.clear()
+        self.clearFields()
 
         self.refreshContacts()
 
@@ -131,3 +186,9 @@ class MainWindow(QMainWindow):
 
     def quit(self):
         QApplication.instance().quit()
+
+    def clearFields(self):
+        self.name_edit.clear()
+        self.email_edit.clear()
+        self.address_edit.clear()
+        self.notes_edit.clear()
